ioctl: nvme_resv_acquire use nvme_passthru_cmd directly
Drop struct nvme_resv_acquire_args.
