+++ test/ioctl/misc.c
@@ -1116,34 +1116,23 @@ static void test_copy(void)
 
 static void test_resv_acquire(void)
 {
-	__u32 result = 0;
-
-	struct nvme_resv_acquire_args args = {
-		.crkey = 0,
-		.nrkey = 0,
-		.result = &result,
-		.args_size = sizeof(args),
-		.nsid = TEST_NSID,
-		.rtype = NVME_RESERVATION_RTYPE_EAAR,
-		.racqa = NVME_RESERVATION_RACQA_PREEMPT,
-		.iekey = true,
-	};
-
+	enum nvme_resv_rtype rtype = NVME_RESERVATION_RTYPE_EAAR;
+	enum nvme_resv_racqa racqa = NVME_RESERVATION_RACQA_PREEMPT;
 	__le64 payload[2] = { 0 };
+	bool iekey = true;
+	__u32 result = 0;
+	int err;
 
 	struct mock_cmd mock_io_cmd = {
 		.opcode = nvme_cmd_resv_acquire,
 		.nsid = TEST_NSID,
-		.cdw10 = (args.racqa & 0x7) | (args.iekey ? 1 << 3 : 0) |
-			 (args.rtype << 8),
+		.cdw10 = (racqa & 0x7) | (iekey ? 1 << 3 : 0) | (rtype << 8),
 		.data_len = sizeof(payload),
 		.in_data = payload,
 	};
 
-	int err;
-
 	set_mock_io_cmds(&mock_io_cmd, 1);
-	err = nvme_resv_acquire(test_link, &args);
+	err = nvme_resv_acquire(test_link, TEST_NSID, 0, 0, rtype, racqa, iekey, &result);
 	end_mock_cmds();
 	check(err == 0, "returned error %d", err);
 	check(result == 0, "returned result %u", result);
